refactor(#3096): split arrow_convert.rs by responsibility (no behavior change)
